requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
openpyxl==3.1.5
tqdm==4.67.1
datasets=3.0.1
